Used szymon improved common command
